Continued x4 porting
