--- src/main/java/team/gif/robot/commands/climber/ControlPanelDown.java
@@ -1,47 +1,0 @@
-package team.gif.robot.commands.climber;
-
-import edu.wpi.first.wpilibj2.command.CommandBase;
-import team.gif.robot.Constants;
-import team.gif.robot.Globals;
-import team.gif.robot.Robot;
-
-public class ControlPanelDown extends CommandBase {
-    @SuppressWarnings({"PMD.UnusedPrivateField", "PMD.SingularField"})
-
-    private static boolean finished = false;
-
-    public ControlPanelDown() {
-        // Use addRequirements() here to declare subsystem dependencies.
-    }
-
-    // Called when the command is initially scheduled.
-    @Override
-    public void initialize() {
-        finished = false;
-        Globals.climberMotorEnabled = true;
-    }
-
-    // Called every time the scheduler runs while the command is scheduled.
-    @Override
-    public void execute() {
-        if (Robot.climber.getPosition() > Constants.Climber.MIN_POS) {
-            Robot.climber.setSpeed(-0.6);
-        } else {
-            finished = true;
-        }
-    }
-
-
-    // Returns true when the command should end.
-    @Override
-    public boolean isFinished() {
-        return finished;
-    }
-
-    // Called once the command ends or is interrupted.
-    @Override
-    public void end(boolean interrupted) {
-        Robot.climber.setSpeed(0);
-        Globals.climberMotorEnabled = false;
-    }
-}
